+++ components/common/header/List.tsx
@@ -5,7 +5,7 @@ import toggleIcon from '@/assets/image/icon/header-toggle.svg';
 import TeamList from './TeamList';
 import { useUserStore } from '@/store/authStore';
 import { useGroupStore } from '@/store/useGroupStore';
-import { useUsersQuery } from '@/queries/user/user';
+import { useUsersQuery } from '@/queries/user';
 import { useState, useEffect, useRef } from 'react';
 import useClickOutside from '@/hooks/useClickOutside';
 import { motion } from 'framer-motion';
